+++ src/components/ScheduleTab.jsx
@@ -1,8 +1,8 @@
 // @flow
 import React, { useMemo } from "react";
 import Breadcrumbs from "./Breadcrumbs";
-import { uid, saveDB } from "../App";
-import type { DB, ScheduleSlot } from "../App";
+import { uid, saveDB } from "../state/appState";
+import type { DB, ScheduleSlot } from "../types";
 
 export default function ScheduleTab({ db, setDB }: { db: DB; setDB: (db: DB) => void }) {
   const byArea = useMemo(() => {
