+++ src/components/TasksTab.jsx
@@ -2,8 +2,8 @@
 import React, { useState } from "react";
 import Breadcrumbs from "./Breadcrumbs";
 import Modal from "./Modal";
-import { fmtDate, uid, saveDB, todayISO } from "../App";
-import type { DB, TaskItem } from "../App";
+import { fmtDate, uid, saveDB, todayISO } from "../state/appState";
+import type { DB, TaskItem } from "../types";
 
 export default function TasksTab({ db, setDB }: { db: DB; setDB: (db: DB) => void }) {
   const [edit, setEdit] = useState<TaskItem | null>(null);
