+++ src/components/__tests__/ScheduleTab.groups.test.jsx
@@ -4,7 +4,7 @@ import { render, screen, within, cleanup } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import "@testing-library/jest-dom";
 
-jest.mock("../../App", () => ({
+jest.mock("../../state/appState", () => ({
   uid: () => "test-id",
   saveDB: jest.fn(),
   todayISO: () => new Date().toISOString(),
